added chat page URL in dashboard

+++ pollsite/poll/consumers.py
@@ -498,8 +498,9 @@ class MeetingConsumer(WebsocketConsumer):
 		self.ytHandler._polling = True
 
 	def terminate_yt_polling(self):
-		self.ytHandler.terminate()
-		self.ytHandler = None
+		if(self.ytHandler):
+			self.ytHandler.terminate()
+			self.ytHandler = None
 
 	def stop_yt_polling(self,question):
 		if(not self.ytHandler):
@@ -525,8 +526,9 @@ class MeetingConsumer(WebsocketConsumer):
 		self.twHandler.run()
 
 	def terminate_tw_polling(self):
-		self.twHandler.terminate()
-		self.twHandler = None
+		if(self.twHandler):
+			self.twHandler.terminate()
+			self.twHandler = None
 		# self.twHandler = None
 
 	def stop_tw_polling(self,question):
